🐛 📌 Pin `@storybook/router` to v6.4.x
Something in `@storybook/router`s v6.5 build breaks with ESM modules. Since
we're going to be getting rid of this dependency in a future PR very soon,
rather than sinking time into figuring out what they broke (or what we might
doing wrong), it'll be easier to pin the dependency to a working version until
we remove the dependency outright.